Fixed better config system
Added a better config system for loading configuration files.

Regex for loading files can be replaced by custom regex and the system has support for any setting not like before when it was static.

Due to non static settings, settings can be null and cause errors but they are not static so any settings are supported.